Moved pause to it's own file.  Corrected documentation on attachment.
